web-animations/animation-player/animation-player-source-change.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player change source attribute tests</title>
<meta name="assert" content="A player can only be associated with at most one animation node, and likewise, an animation node can only be associated with at most one player.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-source">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);

    player.source = null;

    assert_equals(player.source, null, 'Value of AnimationPlayer.source attribute ' +
        'should be set to null');
    assert_equals(animation.player, null, 'Old value of AnimationPlayer.source should ' +
        'be disassociated from the player');
}, 'Test setting AnimationPlayer.source attribute to null');

test(function() {
    var player = document.timeline.play(null);
    var animation = newAnimation(createDiv(this));
    player.source = animation;

    assert_equals(player.source, animation, 'Value of AnimationPlayer.source attribute ' +
        'should be set');
    assert_equals(animation.player, player, 'Value of AnimationNode.player should ' +
        'be set to associated player');
}, 'Test setting AnimationPlayer.source attribute to non null value. Initial value ' +
    'of the source attribute is null');

test(function() {
    var animation = newAnimation(createDiv(this));
    var player = document.timeline.play(animation);
    player.source = animation;

    assert_equals(player.source, animation, 'Value of AnimationPlayer.source attribute ' +
        'should not be changed');
    assert_equals(animation.player, player, 'Value of AnimationNode.player should ' +
        'not be changed');
}, 'Test setting AnimationPlayer.source attribute to the same value');

test(function() {
    var animation1 = newAnimation(createDiv(this));
    var player = document.timeline.play(animation1);

    var animation2 = newAnimation(createDiv(this));
    player.source = animation2;

    assert_equals(player.source, animation2, 'Value of AnimationPlayer.source attribute ' +
        'should be set');
    assert_equals(animation2.player, player, 'Value of AnimationNode.player should ' +
        'be set to associated player');
    assert_equals(animation1.player, null, 'Old value of AnimationPlayer.source should ' +
        'be disassociated from the player');
}, 'Test changing the AnimationPlayer.source attribute. Initial value ' +
    'of the source attribute is not null');

test(function() {
    var animation1 = newAnimation(createDiv(this));
    var player1 = document.timeline.play(animation1);
    var animation2 = newAnimation(createDiv(this));
    var player2 = document.timeline.play(animation2);

    player2.source = animation1;

    assert_equals(player2.source, animation1, 'Value of AnimationPlayer.source attribute ' +
        'should be changed');
    assert_equals(animation1.player, player2, 'Value of AnimationNode.player should ' +
        'be set to associated player');
    assert_equals(animation2.player, null, 'Old value of AnimationPlayer.source should ' +
        'be disassociated from the player');
    assert_equals(player1.source, null, 'New value of AnimationPlayer.source attribute ' +
        'should be disassociated from its previous player');
}, 'Test that value of AnimationPlayer.source attribute is disassociated from ' +
    'its previous player');

test(function() {
    var animation1 = newAnimation(createDiv(this));
    var animation2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([animation1, animation2]);
    var animation3 = newAnimation(createDiv(this));
    var player = document.timeline.play(animation3);

    player.source = animation1;

    assert_equals(player.source, animation1, 'Value of AnimationPlayer.source attribute ' +
        'should be set to new value');
    assert_equals(animation1.player, player, 'Value of AnimationNode.player should ' +
        'be set to associated player');
    assert_equals(animation3.player, null, 'Old value of AnimationPlayer.source should ' +
        'be disassociated from the player');
    assert_equals(animation1.parent, null, 'New value of AnimationPlayer.source attribute ' +
        'should be removed from its parent group');
    assert_array_equals(animationGroup.children, [animation2], 'New value of ' +
        'AnimationPlayer.source attribute should be removed from the parent group');
}, 'Test that value of AnimationPlayer.source attribute is removed from ' +
    'its parent group');

test(function() {
    var animation1 = newAnimation(createDiv(this));
    var animation2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([animation1, animation2]);

    var player = document.timeline.play(null);

    player.source = animationGroup;

    assert_equals(player.source, animationGroup, 'Value of AnimationPlayer.source attribute ' +
        'should be set');
    assert_equals(animationGroup.player, player, 'AnimationGroup should ' +
        'be associated with a player');
    assert_array_equals(animationGroup.children, [animation1, animation2],
        'AnimationGroup children should not be changed');
    assert_equals(animation1.player, player, 'Child AnimationNode should ' +
        'be associated with a player');
    assert_equals(animation2.player, player, 'Second child AnimationNode should ' +
        'be associated with a player');
}, 'Test setting AnimationPlayer.source attribute to AnimationGroup. Initial value ' +
    'of the source attribute is null');


test(function() {
    var animation1 = newAnimation(createDiv(this));
    var animation2 = newAnimation(createDiv(this));
    var animation3 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([animation1, animation2]);

    var player = document.timeline.play(animation3);

    player.source = animationGroup;

    assert_equals(player.source, animationGroup, 'Value of AnimationPlayer.source attribute ' +
        'should be set');
    assert_equals(animationGroup.player, player, 'AnimationGroup should ' +
        'be associated with a player');
    assert_array_equals(animationGroup.children, [animation1, animation2],
        'AnimationGroup children should not be changed');
    assert_equals(animation1.player, player, 'Child AnimationNode should ' +
        'be associated with a player');
    assert_equals(animation2.player, player, 'Second child AnimationNode should ' +
        'be associated with a player');
    assert_equals(animation3.player, null, 'Old value of AnimationPlayer.source should ' +
        'be disassociated with a player');
}, 'Test setting AnimationPlayer.source attribute to AnimationGroup. Initial value ' +
    'of the source attribute is not null');

test(function() {
    var animation1 = newAnimation(createDiv(this));
    var animation2 = newAnimation(createDiv(this));
    var animationGroup1 = new AnimationGroup([animation1, animation2]);
    var animationGroup2 = new AnimationGroup([animationGroup1]);

    var player = document.timeline.play(null);

    player.source = animationGroup1;

    assert_equals(player.source, animationGroup1, 'Value of AnimationPlayer.source attribute ' +
        'should be set');
    assert_equals(animationGroup1.player, player, 'AnimationGroup should ' +
        'be associated with a player');
    assert_array_equals(animationGroup1.children, [animation1, animation2],
        'AnimationGroup children should not be changed');
    assert_equals(animation1.player, player, 'Child AnimationNode should ' +
        'be associated with a player');
    assert_equals(animation2.player, player, 'Second child AnimationNode should ' +
        'be associated with a player');
    assert_equals(animationGroup1.parent, null, 'AnimationGroup should be removed ' +
        'from its parent group');
    assert_array_equals(animationGroup2.children, [],
        'AnimationGroup should be removed from the parent group');
}, 'Test that setting AnimationPlayer.source attribute to AnimationGroup removes it ' +
    'from group\'s parent');
</script>
</body>
